+++ src/main/java/edu/clemson/cs/rsrg/vcgeneration/proofrules/statement/WhileStmtRule.java
@@ -26,6 +26,7 @@ import edu.clemson.cs.rsrg.typeandpopulate.typereasoning.TypeGraph;
 import edu.clemson.cs.rsrg.vcgeneration.proofrules.AbstractProofRuleApplication;
 import edu.clemson.cs.rsrg.vcgeneration.proofrules.ProofRuleApplication;
 import edu.clemson.cs.rsrg.vcgeneration.utilities.*;
+import edu.clemson.cs.rsrg.vcgeneration.utilities.helperstmts.VCConfirmStmt;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
